Add files via upload
Theres some story with the intro and outro functions but if you run them you can comment them out. I missed out on a few bugs such as not accounting for leaving the matrix and going out of bounds. As well as some small stamina bugs. Not outputting your out of ambrosia, but for the most part program works and I like the result. This was for my Greek mythology class. I wanted to do something unique.